+++ src/solver_multitask/mod.rs
@@ -186,9 +186,9 @@ pub fn anderson_accel<T, D, P>(
                         for &j in ws {
                             for idx in X_sparse.indptr[j]..X_sparse.indptr[j + 1] {
                                 for t in 0..n_tasks {
-                                    XW_acc[[X_sparse.indices[idx], t]] = XW_acc
-                                        [[X_sparse.indices[idx], t]]
-                                        + X_sparse.data[idx] * W_acc[[j, t]];
+                                    XW_acc[[X_sparse.indices[idx as usize] as usize, t]] = XW_acc
+                                        [[X_sparse.indices[idx as usize] as usize, t]]
+                                        + X_sparse.data[idx as usize] * W_acc[[j, t]];
                                 }
                             }
                         }
@@ -303,7 +303,8 @@ pub fn bcd_epoch_sparse<T: 'static + Float, D: DatafitMultiTask<T>, P: PenaltyMu
         if sum_diff != T::zero() {
             for i in X.indptr[j]..X.indptr[j + 1] {
                 for t in 0..n_tasks {
-                    XW[[X.indices[i], t]] = XW[[X.indices[i], t]] + diff[t] * X.data[i];
+                    XW[[X.indices[i as usize] as usize, t]] =
+                        XW[[X.indices[i as usize] as usize, t]] + diff[t] * X.data[i as usize];
                 }
             }
         }
